Enable passing config-dir and debug in a virtualenv
Using tox.ini allows us to create a contained environment.  However, the
script still needs to access the secret keyring (so that the user only
needs to authorise once).  Thus this also passes
DBUS_SESSION_BUS_ADDRESS as part of the tox environment.

cherry-picked // had to modify the requirements.txt and tox.ini