Change logger calls to be lazy in formatting
Although Python3 gives us the brilliant f"" formatter, it tends to
be inefficient to use it in logging calls as the string format will take
place regardless of whether the logger actually logs to anywhere.  By
making the formatting lazy, if the log doesn't take place then the
string doesn't get formatted.  Looks a bit Python 2, but is more
efficient.