Add a group_channels() to LaunchpadTools
This function groups channels (the 'risk' part) to the same tracks, as
the launchpad API won't create a recipe that pushes to different tracks;
only to the same track with different risks and/or branches.  We aren't
using branches (yet), so this just groups risks against tracks.  The
function returns a dict of track -> [track/risk,..] so that the
consuming function can create multiple recipes based on tracks for a
single branch configuration.