Add `check-builds` subcommand.
This patch introduces a new subcommand that gets all the builds for the
charm recipes defined, the output can be filtered by architecture and
channel, when requested to `--detect-error` the build log is downloaded
and grep for 'ERROR'.

Usage example:

charmhub-lp-tool -c nova-compute check-builds --arch s390x --detect-error

This command will retrieve all the builds of the nova-compute charm for
the s390x architecture and due to the `--detect-error` it will download
the build log and look for errors.