Cache charm recipes in memory.
This change introduces the use of `functools.cache` to cache the list of
charm recipes, the cache is cleared when a new recipe is created via
`create_charm_recipe()`.

An optimization included in this change is that when filtering the list
of recipes the project comparison is made using the *_link properties
which avoids a trip to Launchpad API to load the project objects.