Optimize iteration of builds per recipe.
This change allows generating a report for openstack, ceph and ovn
across all their tracks and risks in ~30m, before this change the same
report takes ~90m.

This change breaks the loop of over the list of builds associated to a
recipe when the revision id changes (git commit id), since we only want
the last set of builds triggered by the most recent commit.