Add commands to help with managing channels/tracks
The additional commands are to help with managing tracks and risks for
the charms on charmhub.  The commands are:

 - clean-channel
 - change-risk
 - repair-resources

The 'clean-channel' command cleans a channel so that it only retains
revisions for a particular unique set of bases.

The 'change-risk' command enables the promotion of risks from on risk to
another.

The 'repair-resources' command is used to add resources for revisions
that have been released to a channel that were released without
resources when the charmhub allowed that feature.  It's unlikely to be
used more than once, and only for old charms that were released prior to
the enforcement of resources by the charmhub, or that were imported into
the charmhub from the charmstore.

There are also a number of bug fixes, updates to commands and general
usability improvements.