Move defaults from CharmProject._add_branches() to schema.
Until now the schema definition was being used to validate the
configuration passed was syntactically and semantically valid, although
default values defined in it were not being passed into the program.

This change captures the data returned by Schema.validate() and it's
used to instantiate the CharmProject classes, this allows to schema to
own the default values used by this program.