Move upload error message to build log column.
There are scenarios where Launchpad could have uploaded the built charm
and a revision number assigned, but it failed to be released, the most
common reason to fail to release is the missing resource attached to the
Ubuntu series set the charm runs on.

This change makes sure the store revision is included in the report and
the upload error message displayed in the "build log" column.